# infra/cdk/openapi.yaml
openapi: 3.0.3
info:
  title: DeepCopy API
  version: 1.0.0
  description: |
    API for submitting AI processing jobs and retrieving their status and results.

    Authentication (OAuth2 client credentials via Amazon Cognito):

    ```bash
    # Set endpoints (from CDK outputs)
    API_URL="https://o5egokjpsl.execute-api.eu-west-1.amazonaws.com/prod/"
    TOKEN_ENDPOINT="https://deepcopy-613663743323-eu-west-1.auth.eu-west-1.amazoncognito.com/oauth2/token"

    # Your app client credentials (retrieve secret from Cognito console)
    CLIENT_ID="<your-client-id>"
    CLIENT_SECRET="<your-client-secret>"

    # Request access token with read+write scopes
    ACCESS_TOKEN=$(curl -s -u "$CLIENT_ID:$CLIENT_SECRET" \
      -d "grant_type=client_credentials&scope=https://deep-copy.api/read https://deep-copy.api/write" \
      "$TOKEN_ENDPOINT" | jq -r .access_token)

    echo "Token acquired: ${#ACCESS_TOKEN} chars"

    # Submit a v2 job
    curl -sS -X POST \
      -H "Authorization: Bearer $ACCESS_TOKEN" \
      -H "content-type: application/json" \
      -d '{
        "sales_page_url": "https://try.meritrelief.com/",
        "project_name": "meritrelief",
        "advertorial_type": "Listicle",
        "gender": "Female"
      }' \
      "${API_URL}v2/jobs"

    # Get v2 job status (replace with real jobId)
    curl -sS -H "Authorization: Bearer $ACCESS_TOKEN" "${API_URL}v2/jobs/YOUR_JOB_ID"

    # Get v2 job result JSON (404 until available)
...
    ```
servers:
...
security:
  - cognitoOAuth2: [ ]
paths:
  /jobs:
    options:
      summary: CORS preflight
      responses:
        '204':
          description: No Content
          headers:
            Access-Control-Allow-Origin: { $ref: '#/components/headers/AccessControlAllowOrigin' }
            Access-Control-Allow-Headers: { $ref: '#/components/headers/AccessControlAllowHeaders' }
            Access-Control-Allow-Methods: { $ref: '#/components/headers/AccessControlAllowMethods' }
    post:
      summary: Submit a new job
      operationId: submitJob
      security:
        - cognitoOAuth2:
            - https://deep-copy.api/write
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
            examples:
              minimal:
                summary: Minimal payload
                value: {}
              full:
                summary: Full payload
                value:
                  sales_page_url: "https://try.meritrelief.com/"
                  project_name: "meritrelief"
                  advertorial_type: "Listicle"
                  customer_avatars:
                    - persona_name: "Active Senior"
                      characteristics:
                        - "Active lifestyle"
                        - "Gardening enthusiast"
                        - "Chronic pain"
                      description: "Retired professional who enjoys gardening and light exercise but struggles with chronic sciatic pain"
                      age_range: "60-75"
                      gender: "both"
                      key_buying_motivation: "Seeking non-invasive pain relief to maintain active lifestyle"
                      pain_point: "Chronic sciatic pain prevents them from enjoying favorite activities like gardening"
                      emotion: "Frustrated and limited, feeling their independence slipping away"
                      desire: "A natural solution that allows them to stay active without relying on prescription painkillers"
                      objections:
                        - "Will this actually work for chronic pain?"
                        - "Is it safe for someone my age?"
                        - "Can I afford another wellness product?"
                      failed_alternatives:
                        - "Prescription pain medication with unwanted side effects"
                        - "Physical therapy that didn't provide lasting relief"
                        - "Over-the-counter pain relievers that only work temporarily"
                    - persona_name: "Desk-Bound Professional"
                      characteristics:
                        - "Standing all day"
                        - "Work-related pain"
                        - "Evening fatigue"
                      description: "Office worker sitting 8+ hours daily experiencing lower back pain"
                      age_range: "30-45"
                      gender: "both"
                      key_buying_motivation: "Quick relief from work-related pain"
                      pain_point: "Stands all day at work (nurses, retail staff, warehouse workers). Feet ache, swell, and feel heavy every evening."
                      emotion: "Feels drained and defeated when coming home, wants relief to enjoy family time instead of collapsing on the couch."
                      desire: "Quick daily solution that soothes feet and restores energy without expensive spa visits."
                      objections:
                        - "Will this actually help after a long shift?"
                        - "Can I afford another wellness expense?"
                        - "Do I have time to use this daily?"
                      failed_alternatives:
                        - "Compression socks that don't provide enough relief"
                        - "Soaking feet in warm water (temporary relief only)"
                        - "Pain medication with unwanted side effects"
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitJobResponse'
              examples:
                accepted:
                  value:
                    jobId: "c3f0e0c4-3a61-4c52-8b5a-2e0a8b8b4a5f"
                    taskArn: "arn:aws:ecs:eu-west-1:123456789012:task/cluster/1234567890abcdef"
                    status: "SUBMITTED"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codeSamples:
        - lang: bash
          label: curl (submit)
          source: |-
            API_URL="https://o5egokjpsl.execute-api.eu-west-1.amazonaws.com/prod/"
            curl -sS -X POST \
              -H "Authorization: Bearer $ACCESS_TOKEN" \
              -H "content-type: application/json" \
              -d '{
                "sales_page_url": "https://try.meritrelief.com/",
                "project_name": "meritrelief",
                "advertorial_type": "Listicle",
                "customer_avatars": [
                  {
                    "persona_name": "Active Senior",
                    "characteristics": ["Active lifestyle", "Gardening enthusiast", "Chronic pain"],
                    "description": "Retired professional who enjoys gardening and light exercise but struggles with chronic sciatic pain",
                    "age_range": "60-75",
                    "gender": "both",
                    "key_buying_motivation": "Seeking non-invasive pain relief to maintain active lifestyle"
                  }
                ]
              }' \
              "${API_URL}jobs"
...
  /jobs/{id}/result:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    options:
      summary: CORS preflight
      responses:
        '204':
          description: No Content
          headers:
            Access-Control-Allow-Origin: { $ref: '#/components/headers/AccessControlAllowOrigin' }
            Access-Control-Allow-Headers: { $ref: '#/components/headers/AccessControlAllowHeaders' }
            Access-Control-Allow-Methods: { $ref: '#/components/headers/AccessControlAllowMethods' }
    get:
      summary: Get job result
      operationId: getJobResult
      security:
        - cognitoOAuth2:
            - https://deep-copy.api/read
      responses:
        '200':
          description: Job result JSON or text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResult'
            text/plain:
              schema:
                type: string
        '404':
          description: Result not available
          content:
            text/plain:
              schema:
                type: string
                example: "Result not available"
        '500':
          description: Server error
          content:
            text/plain:
              schema:
                type: string
      x-codeSamples:
        - lang: bash
          label: curl (result)
          source: |-
            API_URL="https://o5egokjpsl.execute-api.eu-west-1.amazonaws.com/prod/"
            curl -sS -H "Authorization: Bearer $ACCESS_TOKEN" "${API_URL}jobs/YOUR_JOB_ID/result"
...
  /v2/jobs:
    options:
      summary: CORS preflight
      responses:
        '204':
          description: No Content
          headers:
            Access-Control-Allow-Origin: { $ref: '#/components/headers/AccessControlAllowOrigin' }
            Access-Control-Allow-Headers: { $ref: '#/components/headers/AccessControlAllowHeaders' }
            Access-Control-Allow-Methods: { $ref: '#/components/headers/AccessControlAllowMethods' }
    post:
      summary: Submit a new job (V2 - General Research Flow)
      description: |
        V2 endpoint for submitting jobs using the new general research flow.
        This endpoint performs deep research to automatically identify multiple customer avatars
        and generates marketing angles for each.
        
        **Breaking changes from v1:**
        - `sales_page_url` and `project_name` are now required
        - `customer_avatars`, `persona`, and `age_range` fields are rejected (400 error)
        - Response includes `identified_avatars`, multiple `avatar_sheet` entries, and `necessary_beliefs`
      operationId: submitJobV2
      security:
        - cognitoOAuth2:
            - https://deep-copy.api/write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobV2Request'
            examples:
              minimal:
                summary: Minimal required payload
                value:
                  sales_page_url: "https://try.meritrelief.com/"
                  project_name: "meritrelief"
              full:
                summary: Full payload with optional fields
                value:
                  sales_page_url: "https://try.meritrelief.com/"
                  project_name: "meritrelief"
                  advertorial_type: "Listicle"
                  research_requirements: "Focus on natural ingredients and FDA approval status"
                  gender: "Female"
                  location: "USA"
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitJobResponse'
              examples:
                accepted:
                  value:
                    jobId: "c3f0e0c4-3a61-4c52-8b5a-2e0a8b8b4a5f"
                    status: "SUBMITTED"
        '400':
          description: Invalid request body (missing required fields or deprecated fields used)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                deprecated_fields:
                  summary: Deprecated fields used
                  value:
                    error: "Deprecated fields not allowed in v2 API: customer_avatars"
                missing_required:
                  summary: Missing required fields
                  value:
                    error: "Missing required fields: project_name, sales_page_url"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codeSamples:
        - lang: bash
          label: curl (submit v2)
          source: |-
            API_URL="https://o5egokjpsl.execute-api.eu-west-1.amazonaws.com/prod/"
            curl -sS -X POST \
              -H "Authorization: Bearer $ACCESS_TOKEN" \
              -H "content-type: application/json" \
              -d '{
                "sales_page_url": "https://try.meritrelief.com/",
                "project_name": "meritrelief",
                "research_requirements": "Focus on natural ingredients",
                "gender": "Female",
                "location": "USA"
              }' \
              "${API_URL}v2/jobs"
...
components:
  securitySchemes:
    cognitoOAuth2:
      type: oauth2
      description: |
        OAuth2 client credentials against Amazon Cognito User Pool domain.
        Use the token endpoint output by the stack to obtain an access token.
      flows:
        clientCredentials:
          tokenUrl: https://deepcopy-613663743323-eu-west-1.auth.eu-west-1.amazoncognito.com/oauth2/token
          scopes:
            https://deep-copy.api/read: Read jobs and results
            https://deep-copy.api/write: Submit jobs
  schemas:
    SubmitJobV2Request:
      type: object
      description: Request body for V2 jobs endpoint (general research flow)
      required: [sales_page_url, project_name]
      additionalProperties: false
      properties:
        sales_page_url:
          type: string
          format: uri
          description: URL to the sales page to analyze (required)
        project_name:
          type: string
          description: Name of the project (required)
        advertorial_type:
          type: string
          description: Type of advertorial to generate (e.g., 'Listicle', 'Advertorial')
        research_requirements:
          type: string
          description: Optional specific requirements or focus areas for the research
        gender:
          type: string
          enum: [Male, Female, Both]
          description: Optional gender target
        location:
          type: string
          description: Optional location target
        target_product_name:
          type: string
          description: Optional product name to use consistently across all pipeline outputs
      example:
        sales_page_url: "https://try.meritrelief.com/"
        project_name: "meritrelief"
        advertorial_type: "Listicle"
        research_requirements: "Focus on natural ingredients"
        gender: "Female"
        location: "USA"
...
    SubmitJobRequest:
      type: object
      additionalProperties: true
      properties:
        sales_page_url:
          type: string
          format: uri
          description: URL to the sales page to analyze
        project_name:
          type: string
        advertorial_type:
          type: string
        research_requirements:
          type: string
          description: Optional specific requirements or focus areas for the research
        gender:
          type: string
          enum: [Male, Female, Both]
          description: Optional gender target
        location:
          type: string
          description: Optional location target
        customer_avatars:
          type: array
          description: (Deprecated) List of customer avatars/personas to target. Ignored in new flow.
          items:
            $ref: '#/components/schemas/CustomerAvatar'
        persona:
          type: string
          deprecated: true
          description: (Deprecated) Use customer_avatars instead
        age_range:
          type: string
          deprecated: true
          description: (Deprecated) Use customer_avatars instead
        gender_deprecated:
          type: string
          deprecated: true
          description: (Deprecated) Use 'gender' field instead
      example:
        sales_page_url: "https://try.meritrelief.com/"
        project_name: "meritrelief"
        advertorial_type: "Listicle"
        research_requirements: "Focus on natural ingredients"
        gender: "Female"
        location: "USA"
    SubmitJobResponse:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
          format: uuid
        taskArn:
          type: string
          nullable: true
        status:
          type: string
          enum: [SUBMITTED]
...
    JobResult:
      type: object
      description: Structured JSON produced by the pipeline including rewritten advertorial content.
      additionalProperties: false
      properties:
        project_name:
          type: string
        timestamp_iso:
          type: string
          format: date-time
        job_id:
          type: string
        api_version:
          type: string
          description: API version (v1)
        results:
          type: object
          additionalProperties: false
          properties:
            research_page_analysis:
              type: string
            doc1_analysis:
              type: string
            doc2_analysis:
              type: string
            deep_research_prompt:
              type: string
            deep_research_output:
              type: string
            avatar_sheet:
              $ref: '#/components/schemas/AvatarList'
            identified_avatars:
              $ref: '#/components/schemas/IdentifiedAvatarList'
            offer_brief:
              $ref: '#/components/schemas/OfferBrief'
            marketing_philosophy_analysis:
              type: string
            necessary_beliefs:
              type: string
            summary:
              type: string
            swipe_results:
              type: array
              items:
                $ref: '#/components/schemas/SwipeResult'
            marketing_angles:
              type: array
              items:
                $ref: '#/components/schemas/Angle'
            customer_avatars:
              type: array
              description: Simplified list of identified avatars
              items:
                $ref: '#/components/schemas/IdentifiedAvatar'
      required: [project_name, timestamp_iso, results, job_id]
    SwipeResult:
      type: object
      description: Result item for a specific marketing angle containing structured advertorial content.
      required: [angle, content]
      properties:
        angle:
          type: string
          description: Marketing angle used to generate the content.
        content:
          description: Structured content; either a Listicle or an Avertorial.
          oneOf:
            - $ref: '#/components/schemas/Listicle'
            - $ref: '#/components/schemas/Advertorial'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    Angle:
      type: object
      required: [angle, title, target_age_range, target_audience, pain_points, desires, common_objections, failed_alternatives, copy_approach]
      properties:
        angle:
          type: string
          description: "The marketing angle for the product"
        title:
          type: string
          description: "The title of the marketing angle"
        target_age_range:
          type: string
          description: "The target age range for the marketing angle. Note it like: 45-55, 60-70, etc."
        target_audience:
          type: string
          description: "The target audience for the marketing angle. Note it like: Competitive athletes seeking reliability, Skilled professionals seeking productivity, etc."
        pain_points:
          type: array
          description: "The pain points for the marketing angle"
          items:
            type: string
        desires:
          type: array
          description: "The desires for the marketing angle"
          items:
            type: string
        common_objections:
          type: array
          description: "The common objections for the marketing angle"
          items:
            type: string
        failed_alternatives:
          type: array
          description: "The failed alternatives for the marketing angle"
          items:
            type: string
        copy_approach:
          type: array
          description: "The copy approaches for the marketing angle"
          items:
            type: string
    # ===== New Avatar Enums =====
...
    CustomerAvatar:
      type: object
      required: [persona_name, characteristics, description, age_range, gender, key_buying_motivation, pain_point, emotion, desire, objections, failed_alternatives]
      properties:
        persona_name:
          type: string
          description: A short, descriptive name that captures the essence of the customer
        characteristics:
          type: array
          description: Characteristics of the customer, e.g. Stands all day, Active lifestyle, etc. Provide at least 3 characteristics and max three words each.
          items:
            type: string
        description:
          type: string
          description: 1 sentence summarizing their lifestyle, motivations, and key challenges
        age_range:
          type: string
          description: Approximate age bracket (e.g., 25-34)
        gender:
          type: string
          enum: [Male, Female, Both]
          description: male, female, or both
        key_buying_motivation:
          type: string
          description: What drives them to purchase this product
        pain_point:
          type: string
          description: Specific problem or challenge the customer faces
        emotion:
          type: string
          description: Emotion they feel related to the pain point
        desire:
          type: string
          description: What they desire as a solution to their pain point
        objections:
          type: array
          description: List of objections or concerns they might have
          items:
            type: string
        failed_alternatives:
          type: array
          description: List of alternatives they've tried that didn't work
          items:
            type: string
        is_broad_avatar:
          type: boolean
          description: Whether this avatar represents the overall customer base
    IdentifiedAvatar:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          description: Name of the avatar (e.g., 'Busy Mom')
        description:
          type: string
          description: Brief description of who they are and why they are a good fit
    IdentifiedAvatarList:
      type: object
      required: [avatars]
      properties:
        avatars:
          type: array
          items:
            $ref: '#/components/schemas/IdentifiedAvatar'
          description: List of potential customer avatars
    AvatarList:
      type: object
      required: [avatars]
      properties:
        avatars:
          type: array
          items:
            $ref: '#/components/schemas/Avatar'
          description: List of identified customer avatars
    ConsciousnessLevel:
      type: string
      enum: [low, high]
      description: How self-aware and change-ready the prospect is
    AwarenessLevel:
      type: string
      enum: [unaware, problem aware, solution aware, product aware, most aware]
      description: Prospect's awareness per Schwartz's stages
...
    OfferBrief:
      type: object
      description: Structured brief to guide AI when generating copy, concepts, and funnel assets
      properties:
        market_snapshot:
          $ref: '#/components/schemas/MarketSnapshot'
        big_idea_section:
          $ref: '#/components/schemas/BigIdeaSection'
        product:
          $ref: '#/components/schemas/ProductInfo'
        pain_desire:
          $ref: '#/components/schemas/PainDesireSection'
        failed_solutions:
          $ref: '#/components/schemas/FailedSolutionsSection'
        competitor_landscape:
          $ref: '#/components/schemas/CompetitorLandscape'
        belief_architecture:
          $ref: '#/components/schemas/BeliefArchitecture'
        objections_section:
          $ref: '#/components/schemas/ObjectionsSection'
        research_inspiration:
          $ref: '#/components/schemas/ResearchInspiration'
        # Legacy fields
        potential_product_names:
          type: array
          items:
            type: string
        level_of_consciousness:
          $ref: '#/components/schemas/ConsciousnessLevel'
        level_of_awareness:
          $ref: '#/components/schemas/AwarenessLevel'
        stage_of_sophistication:
          $ref: '#/components/schemas/StageOfSophistication'
        big_idea:
          type: string
          nullable: true
        metaphors:
          type: array
          items:
            type: string
        potential_ump:
          type: array
          items:
            type: string
        potential_ums:
          type: array
          items:
            type: string
        guru:
          type: string
          nullable: true
        discovery_story:
          type: string
          nullable: true
        headline_ideas:
          type: array
          items:
            $ref: '#/components/schemas/HeadlineIdea'
        objections:
          type: array
          items:
            type: string
        belief_chains:
          type: array
          items:
            $ref: '#/components/schemas/BeliefChain'
        funnel_architecture:
          type: array
          items:
            type: string
        potential_domains:
          type: array
          items:
            type: string
        examples_swipes:
          type: array
          items:
            $ref: '#/components/schemas/SwipeExample'
...
  headers:
    AccessControlAllowOrigin:
      schema:
        type: string
      example: '*'
    AccessControlAllowHeaders:
      schema:
        type: string
      example: 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'
    AccessControlAllowMethods:
      schema:
        type: string
      example: 'GET,POST,OPTIONS'